Refactor SmallScenerySetColourAction hpp=>h/cpp
