+++ src/openrct2-ui/windows/TopToolbar.cpp
@@ -39,7 +39,7 @@
 #include <openrct2/actions/PauseToggleAction.h>
 #include <openrct2/actions/SetCheatAction.h>
 #include <openrct2/actions/SmallSceneryPlaceAction.h>
-#include <openrct2/actions/SmallScenerySetColourAction.hpp>
+#include <openrct2/actions/SmallScenerySetColourAction.h>
 #include <openrct2/actions/SurfaceSetStyleAction.hpp>
 #include <openrct2/actions/WallPlaceAction.hpp>
 #include <openrct2/actions/WallSetColourAction.hpp>
